+++ common/src/main/java/net/modgarden/barricade/mixin/FaceAttachedHorizontalDirectionalBlockAccessor.java
@@ -1,0 +1,15 @@
+package net.modgarden.barricade.mixin;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.FaceAttachedHorizontalDirectionalBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(FaceAttachedHorizontalDirectionalBlock.class)
+public interface FaceAttachedHorizontalDirectionalBlockAccessor {
+	@Invoker
+	BlockState invokeUpdateShape(BlockState state, Direction facing, BlockState facingState, LevelAccessor level, BlockPos currentPos, BlockPos facingPos);
+}